Fixing express mocking.
