+++ tests/express-mock.ts
@@ -1,0 +1,46 @@
+// @see https://github.com/swc-project/jest/issues/14#issuecomment-970189585
+import http from "http";
+
+const compressionMock = jest.fn();
+jest.mock("compression", () => compressionMock);
+
+const staticHandler = jest.fn();
+const staticMock = jest.fn(() => staticHandler);
+
+let appMock: ReturnType<typeof newAppMock>;
+const expressJsonMock = jest.fn();
+const newAppMock = () => ({
+  disable: jest.fn(),
+  use: jest.fn(),
+  listen: jest.fn((port, cb) => {
+    if (cb) {
+      cb();
+    }
+    return new http.Server();
+  }),
+  get: jest.fn(),
+  post: jest.fn(),
+  options: jest.fn(),
+});
+
+const renewApp = () => {
+  appMock = newAppMock();
+};
+
+const expressMock = jest.mock("express", () => {
+  renewApp();
+  const returnFunction = () => appMock;
+  returnFunction.json = () => expressJsonMock;
+  returnFunction.static = staticMock;
+  return returnFunction;
+});
+
+export {
+  compressionMock,
+  expressMock,
+  appMock,
+  expressJsonMock,
+  renewApp,
+  staticMock,
+  staticHandler,
+};
