+++ tests/unit/server.spec.ts
@@ -1,33 +1,12 @@
 import http from "http";
-
-let appMock: ReturnType<typeof newAppMock>;
-const expressJsonMock = jest.fn();
-const newAppMock = () => ({
-  disable: jest.fn(),
-  use: jest.fn(),
-  listen: jest.fn((port, cb) => {
-    if (cb) {
-      cb();
-    }
-    return new http.Server();
-  }),
-  get: jest.fn(),
-  post: jest.fn(),
-  options: jest.fn(),
-});
-
-const expressMock = jest.mock("express", () => {
-  appMock = newAppMock();
-  const returnFunction = () => appMock;
-  returnFunction.json = () => expressJsonMock;
-  return returnFunction;
-});
-
-const compressionMock = jest.fn();
-jest.mock("compression", () => compressionMock);
-
-import express, { Request, Response } from "express"; // express is mocked above
 import https from "https";
+import {
+  appMock,
+  compressionMock,
+  expressJsonMock,
+  expressMock,
+  renewApp,
+} from "../express-mock";
 import { Logger } from "winston";
 import { z } from "zod";
 import {
@@ -42,11 +21,10 @@ import {
   createNotFoundHandler,
   createParserFailureHandler,
 } from "../../src/server";
+import express, { Request, Response } from "express"; // express is mocked server-mock.inc.ts
 
 describe("Server", () => {
-  beforeEach(() => {
-    appMock = newAppMock();
-  });
+  beforeEach(renewApp);
 
   afterAll(() => {
     jest.restoreAllMocks();
